requirements.txt
click==7.1.2
numpy==1.19.2
pandas==1.1.3
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
-e git+https://github.com/nightingal3/metaskills-analysis.git@531ac0157020cc4e5cd13ca0e9d3c7fbf2f18671#egg=src
